moved sample client out of nosql land to test transactionality

+++ sample-engine-client/src/main/java/org/cucina/sample/engine/client/ItemRepository.java
@@ -1,7 +1,5 @@
 package org.cucina.sample.engine.client;
 
-import java.math.BigInteger;
-
 import org.springframework.data.repository.CrudRepository;
 
 
@@ -11,5 +9,5 @@ import org.springframework.data.repository.CrudRepository;
  * @author vlevine
  */
 public interface ItemRepository
-    extends CrudRepository<Item, BigInteger> {
+    extends CrudRepository<Item, Long> {
 }
